refactor: address PR review comments for MLE/MLD decoders
- Convert mld.py and mle.py into subpackages with separate modules
  for decoder logic, sinter interface, and utilities
- Move sinter classes to dedicated sinter_interface.py submodules
  so importing the core decoder does not require sinter
- Move helper functions (pack/unpack, shots_to_counts) to mld/utils.py
- Remove STEP_SIZE global constant, make it a constructor argument
- Remove test-only properties (det_obs_counts, is_cached_correction,
  is_cached_df, max_observable_index, detector_vertices, weights,
  observable_indices)
- Inline _get_env into decode_error
- Fix scipy import error message to suggest pip install scipy
- Use enumerate instead of range(len(...)) in decode_error loops
- Remove @overload decorators from GurobiDecoder.decode
- Remove trivial tests that only test stim DEM or default values
- Inline single-use test helpers into their respective tests
- Remove SinterTableDecoder/SinterGurobiDecoder from top-level exports
