+++ src/bloqade/decoders/_decoders/__init__.py
@@ -1,11 +1,5 @@
-from .mld import (
-    TableDecoder as TableDecoder,
-    SinterTableDecoder as SinterTableDecoder,
-)
-from .mle import (
-    GurobiDecoder as GurobiDecoder,
-    SinterGurobiDecoder as SinterGurobiDecoder,
-)
+from .mld import TableDecoder as TableDecoder
+from .mle import GurobiDecoder as GurobiDecoder
 from .base import BaseDecoder as BaseDecoder
 from .ldpc import (
     BpLsdDecoder as BpLsdDecoder,
