+++ src/bloqade/decoders/_decoders/mld/__init__.py
@@ -1,0 +1,1 @@
+from .decoder import TableDecoder as TableDecoder
