+++ test/test_imports.py
@@ -9,8 +9,6 @@ def test_decoder_exports():
     assert hasattr(decoders, "MWPFDecoder")
     assert hasattr(decoders, "GurobiDecoder")
     assert hasattr(decoders, "TableDecoder")
-    assert hasattr(decoders, "SinterGurobiDecoder")
-    assert hasattr(decoders, "SinterTableDecoder")
     assert hasattr(decoders, "dialects")
 
     assert hasattr(decoders, "Detector")
